% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condeptree.R
\name{condeptree}
\alias{condeptree}
\title{This is the function that computes the skeletion tree from data.
The input is a matrix x whose rows are the data vectors.
The sample size n is the number of rows.
The number of variables p is the number of columns
The function outputs the skeleton tree g.}
\usage{
condeptree(x)
}
\arguments{
\item{x}{The input data matrix.}
}
\description{
This is the function that computes the skeletion tree from data.
The input is a matrix x whose rows are the data vectors.
The sample size n is the number of rows.
The number of variables p is the number of columns
The function outputs the skeleton tree g.
}
