% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outgoing.R
\name{outgoing}
\alias{outgoing}
\title{Creates an outgoing tree from a given undirected treee.}
\usage{
outgoing(tree, dir_tree = NULL, a = NULL, b = 1)
}
\arguments{
\item{tree}{Input tree, undirected.}

\item{dir_tree}{Directionalities that must be present.}

\item{a}{The node being inspected.}

\item{b}{The neighbor being inspected.}
}
\description{
Creates an outgoing tree from a given undirected treee.
}
