% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xicorln.R
\name{xicorln}
\alias{xicorln}
\title{This function computes the xi correlation coefficient.}
\usage{
xicorln(xvec, yvec)
}
\arguments{
\item{xvec}{The vector of x values.}

\item{yvec}{The vector of y values.}
}
\description{
This function computes the xi correlation coefficient.
}
